Improved the maximization function for polynomial regression

+++ models/LinearRegressionPrincetonForget.java
@@ -61,7 +61,7 @@ public class LinearRegressionPrincetonForget implements Regression {
         if (a < 1.0/0.3)
             return (0.3 * a - 0.3 * b - 3) / (0.6 * a - 2);
         else if (a > 1.0/0.3 || b > -2.0/0.3) 
-            return Integer.MAX_VALUE;
+            return (double) Integer.MAX_VALUE;
         else 
             return 1;
     }
